1.1 compat

+++ src/de/fraunhofer/fit/photocompass/model/ApplicationModel.java
@@ -135,7 +135,7 @@ public final class ApplicationModel {
      * @return The current minimum age for photos to be displayed. From 0..1.
      */
 	public float getRelativeMinAge() {
-		return _minAge / MAX_MAX_AGE;
+		return (float) _minAge / (float) MAX_MAX_AGE;
 	}
 	
 	public String getFormattedMinAge() {
